add license data, reorganize libraries for easier perusal
