+++ libraries/chef_haproxy_instance.rb
@@ -1,10 +1,31 @@
 #
 # Cookbook Name:: haproxy-ng
-# Resource:: instance
+# Library:: Chef::Resource::HaproxyInstance,
+#           Chef::Provider::HaproxyInstance
 #
+# Author:: Nathan Williams <nath.e.will@gmail.com>
+#
+# Copyright 2015, Nathan Williams
+#
+# Licensed under the Apache License, Version 2.0 (the "License");
+# you may not use this file except in compliance with the License.
+# You may obtain a copy of the License at
+#
+#     http://www.apache.org/licenses/LICENSE-2.0
+#
+# Unless required by applicable law or agreed to in writing, software
+# distributed under the License is distributed on an "AS IS" BASIS,
+# WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+# See the License for the specific language governing permissions and
+# limitations under the License.
+#
+
+require_relative 'haproxy'
 
 class Chef::Resource
   class HaproxyInstance < Chef::Resource
+    include Haproxy
+
     identity_attr :name
 
     def initialize(name, run_context = nil)
@@ -68,14 +89,6 @@ class Chef::Resource
             spec.all? { |p| p.is_a? Chef::Resource::HaproxyProxy }
           end
         }
-      )
-    end
-
-    def verify(arg = nil)
-      set_or_return(
-        :verify, arg,
-        :kind_of => [TrueClass, FalseClass],
-        :default => true
       )
     end
   end
