+++ libraries/chef_haproxy_userlist.rb
@@ -1,6 +1,23 @@
 #
-# Cookbook Name: haproxy-ng
-# Resource:: userlist
+# Cookbook Name:: haproxy-ng
+# Library:: Chef::Resource::HaproxyUserlist,
+#           Chef::Provider::HaproxyUserlist
+#
+# Author:: Nathan Williams <nath.e.will@gmail.com>
+#
+# Copyright 2015, Nathan Williams
+#
+# Licensed under the Apache License, Version 2.0 (the "License");
+# you may not use this file except in compliance with the License.
+# You may obtain a copy of the License at
+#
+#     http://www.apache.org/licenses/LICENSE-2.0
+#
+# Unless required by applicable law or agreed to in writing, software
+# distributed under the License is distributed on an "AS IS" BASIS,
+# WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+# See the License for the specific language governing permissions and
+# limitations under the License.
 #
 
 class Chef::Resource
