+++ libraries/haproxy.rb
@@ -1,0 +1,34 @@
+#
+# Cookbook Name:: haproxy-ng
+# Library:: Haproxy
+#
+# Author:: Nathan Williams <nath.e.will@gmail.com>
+#
+# Copyright 2015, Nathan Williams
+#
+# Licensed under the Apache License, Version 2.0 (the "License");
+# you may not use this file except in compliance with the License.
+# You may obtain a copy of the License at
+#
+#     http://www.apache.org/licenses/LICENSE-2.0
+#
+# Unless required by applicable law or agreed to in writing, software
+# distributed under the License is distributed on an "AS IS" BASIS,
+# WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+# See the License for the specific language governing permissions and
+# limitations under the License.
+#
+
+require_relative 'haproxy_helpers'
+require_relative 'haproxy_instance'
+require_relative 'haproxy_proxy'
+
+module Haproxy
+  def verify(arg = nil)
+    set_or_return(
+      :verify, arg,
+      :kind_of => [TrueClass, FalseClass],
+      :default => true
+    )
+  end
+end
